requirements.txt
requests[socks]
m3u8
tqdm
pycryptodome
beautifulsoup4